Add files via upload
Chatbot commit, 
Chatbot.py contains the ML algorithm, we use a FC Neural net to span the dictionary query with the greatest number of hits+ provide a proper response for output. Also unpacks the search queries/inputs from the user.

intents.json is a JSON file that Chatbot.py uses as a dictionary for search terms and replies.